% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{approved}
\alias{approved}
\title{Approved packages and functions}
\format{
A tibble with 6 rows and 2 variables:
\describe{
\item{function_name}{Name of the function}
\item{library}{Name of the package}
}
}
\usage{
approved
}
\description{
A dataset that stores approved packages and functions for use.
Each row contains a \code{library} and \code{function_name}.
This dataset is used to illustrate the data format to be stored in
the log.rx.approved option.
}
\examples{
logrx::approved

}
\keyword{datasets}
