% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{approved_functions_init}
\alias{approved_functions_init}
\title{Initialises the timber.approved option}
\usage{
approved_functions_init()
}
\value{
Nothing
}
\description{
Defaults to working directory. This should point to the location of the
dataframe storing approved packages and functions.
}
\details{
See ?approved for an example dataframe.
}
