% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_file_path}
\alias{get_file_path}
\title{Gets full path of file being run}
\usage{
get_file_path(file = NA, normalize = TRUE)
}
\arguments{
\item{file}{File path of file being run, optional}

\item{normalize}{If the returned path should be normalized}
}
\value{
full path of file being run
}
\description{
Gets full path of file being run
}
\examples{
get_file_path()

}
