% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_library}
\alias{get_library}
\title{Add libraries to functions}
\usage{
get_library(df)
}
\arguments{
\item{df}{dataframe containing variables \code{function_name} and \code{SYMBOL_PACKAGE}}
}
\value{
tibble that includes \code{library}
}
\description{
Each script should be independent so we can use the search path since this
would be just for this script.
This must also be run after script execution.
}
