% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_lint_results}
\alias{get_lint_results}
\title{Get lint results}
\usage{
get_lint_results(file)
}
\arguments{
\item{file}{File path of file being run}
}
\value{
results from \code{lintr::lint()}
}
\description{
Pass linters specified in the \code{logrx.lint} option to \code{lintr::lint}
}
