% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_logrx_metadata}
\alias{get_logrx_metadata}
\title{Returns named list of logrx metadata attributes}
\usage{
get_logrx_metadata()
}
\value{
Named list of logrx package metadata attributes
}
\description{
Returns named list of logrx metadata attributes
}
\examples{
get_logrx_metadata()

}
