% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_masked_functions}
\alias{get_masked_functions}
\title{Returns named list of masked functions}
\usage{
get_masked_functions()
}
\value{
Named list of masked functions, source package, and what they mask
}
\description{
Returns named list of masked functions
}
\examples{
get_masked_functions()

}
