% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_session_info}
\alias{get_session_info}
\title{Returns Session Info}
\usage{
get_session_info()
}
\value{
Formatted Session Info
}
\description{
Returns Session Info
}
\examples{
\dontrun{
get_session_info()
}

}
