% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_timber_metadata}
\alias{get_timber_metadata}
\title{Returns named list of timber metadata attributes}
\usage{
get_timber_metadata()
}
\value{
Named list of timber package metadata attributes
}
\description{
Returns named list of timber metadata attributes
}
\examples{
get_timber_metadata()

}
