% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_unapproved_use}
\alias{get_unapproved_use}
\title{Get unapproved packages and functions used}
\usage{
get_unapproved_use(approved_packages, used_packages)
}
\arguments{
\item{approved_packages}{dataframe containing variables \code{function_name} and \code{library}}

\item{used_packages}{dataframe containing variables \code{function_name} and \code{library}}
}
\value{
tibble that includes packages and functions used, but not approved
}
\description{
Compare two dataframes that contain approved and used packages and functions.
}
