% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_cleanup}
\alias{log_cleanup}
\title{Cleans up log and does checks against elements
TODO add end time setting and runtime calculation}
\usage{
log_cleanup()
}
\value{
List of non-NA elements in timber.log environment
}
\description{
Cleans up log and does checks against elements
TODO add end time setting and runtime calculation
}
\examples{
log_cleanup()

}
