% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_init}
\alias{log_init}
\alias{log_config}
\alias{log_cleanup}
\alias{log_write}
\alias{log_remove}
\title{Initialises the timber.log environment}
\usage{
log_init()

log_config(file = NA)

log_cleanup()

log_write()

log_remove()
}
\arguments{
\item{file}{File path of file being run, optional}
}
\value{
Nothing

Nothing

List of non-NA elements and their value in timber.log environment

Nothing

Nothing
}
\description{
Initialises the timber.log environment

Configures the timber.log environment

Cleans up log and does checks against elements

Write the formatted timber.log to a file

Remove the timber.log environment by setting options("timber.log") to NULL
}
\examples{
log_remove()

log_init()

log_config()

log_cleanup()

}
