% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_init}
\alias{log_init}
\title{Initialises the timber.log environment}
\usage{
log_init()
}
\value{
Nothing
}
\description{
Initialises the timber.log environment
}
\examples{
scriptPath <- tempfile()
logDir <- tempdir()
writeLines("print('hello timber')", con = scriptPath)

log_remove()

log_init()

log_config(scriptPath)

log_cleanup()

log_write()

}
