% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_remove}
\alias{log_remove}
\title{Remove the timber.log environment by setting options("timber.log") to the
empty environment}
\usage{
log_remove()
}
\value{
Nothing
}
\description{
Remove the timber.log environment by setting options("timber.log") to the
empty environment
}
\examples{
log_remove()

}
