% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_write}
\alias{log_write}
\title{Write the formatted timber.log to a file}
\usage{
log_write(remove_log_object = TRUE)
}
\arguments{
\item{remove_log_object}{Should the log object be removed after writing, defaults to TRUE}
}
\value{
Nothing
}
\description{
Write the formatted timber.log to a file
}
