% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-execution.R
\name{make_temp_exec_file}
\alias{make_temp_exec_file}
\title{Creation of a execution of a file}
\usage{
make_temp_exec_file(
  file,
  log_name = NA,
  log_path = NA,
  include_rds = FALSE,
  quit_on_error = TRUE,
  to_report = c("messages", "output", "result"),
  show_repo_url = FALSE,
  extra_info = NA
)
}
\arguments{
\item{file}{String. Path to file to execute}

\item{log_name}{String. Name of log file}

\item{log_path}{String. Path to log file}

\item{include_rds}{Boolean. Option to export log object as Rds file.
Defaults to FALSE}

\item{quit_on_error}{Boolean. Should the session quit with status 1 on error?
Defaults to TRUE}

\item{to_report}{String vector. Objects to optionally report, may include as
many as necessary:
\itemize{
\item messages: any messages generated by program execution
\item output: any output generated by program execution
\item result: any result generated by program execution
}}

\item{show_repo_url}{Boolean. Should the repository URLs be reported
Defaults to FALSE}

\item{extra_info}{List. Objects to optionally add on to end of log
in a special extra info section. List printed in YAML format. Optional}
}
\value{
name of tempfile created for running
}
\description{
Creation of a execution of a file
}
