% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_log_file.R
\name{read_log_file}
\alias{read_log_file}
\title{Read and parse previous logrx file}
\usage{
read_log_file(file)
}
\arguments{
\item{file}{String. Path to a logrx log file}
}
\value{
Tibble. Object that includes nested and parsed content
}
\description{
Read and parse previous logrx file
}
\examples{
\dontrun{
read_log_file(previous_log_filepath)
}

}
