% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{run_safely_loudly}
\alias{run_safely_loudly}
\title{Safely run an R script and record results, outputs, messages, errors, warnings}
\usage{
run_safely_loudly(file)
}
\arguments{
\item{file}{File to run}
}
\value{
Nothing
}
\description{
Safely run an R script and record results, outputs, messages, errors, warnings
}
