% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{set_log_name_path}
\alias{set_log_name_path}
\title{Set the log name and path}
\usage{
set_log_name_path(log_name = NA, log_path = NA)
}
\arguments{
\item{log_name}{The log name}

\item{log_path}{The log path}
}
\value{
Nothing
}
\description{
\enumerate{
\item As the name and path if supplied
\item As the file name with .log extension and path if specified or
if they can be determined by the function
\item As logrx_log.log and . if none of the above are specified
}
}
