% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{sink_log_stream}
\alias{sink_log_stream}
\title{Sink the log stream output to a tempfile and write to element}
\usage{
sink_log_stream(file)
}
\arguments{
\item{file}{File to run}
}
\value{
Nothing
}
\description{
Sink the log stream output to a tempfile and write to element
}
