% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timberAddin.R
\name{timberAddin}
\alias{timberAddin}
\title{timberAddin
Code needed to create the miniUI for the Addin}
\usage{
timberAddin()
}
\value{
returns miniUI Addin to batch submit r files and create logs around them
}
\description{
timberAddin
Code needed to create the miniUI for the Addin
}
