% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_errors}
\alias{write_errors}
\title{Format errors attribute for writing}
\usage{
write_errors()
}
\value{
A formatted vector of errors
}
\description{
Format errors attribute for writing
}
\examples{
write_errors()

}
