% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_file_name_path}
\alias{write_file_name_path}
\title{Format file name and path for writing}
\usage{
write_file_name_path()
}
\value{
A vector of file name and path prefixed
}
\description{
Format file name and path for writing
}
