% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_log_element}
\alias{write_log_element}
\title{Generic log element writer}
\usage{
write_log_element(el_key, prefix)
}
\arguments{
\item{el_key}{the key of the element in timber.log to be fetched}

\item{prefix}{string to be placed before element value during formatting}
}
\value{
formatted element including prefix
}
\description{
Generic log element writer
}
\examples{
write_log_element(user, "user running program: ")

}
