% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_log_header}
\alias{write_log_header}
\title{Formatting of log file section headers}
\usage{
write_log_header(title_string)
}
\arguments{
\item{title_string}{String. Used as section title}
}
\value{
Vector of strings. Formatted log file section header
}
\description{
\code{write_log_header} formats a string and returns it as a formatted log file
section header
}
\examples{
\dontrun{
write_log_header("Section Header")
}
}
