% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_log_stream}
\alias{write_log_stream}
\title{Format log_stream attribute for writing}
\usage{
write_log_stream()
}
\value{
A formatted vector of the log stream
}
\description{
Format log_stream attribute for writing
}
