% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_masked_functions}
\alias{write_masked_functions}
\title{Format masked functions attribute for writing}
\usage{
write_masked_functions()
}
\value{
A formatted vector of masked functions
}
\description{
Format masked functions attribute for writing
}
\examples{
write_masked_functions()

}
