% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_messages}
\alias{write_messages}
\title{Format messages attribute for writing}
\usage{
write_messages()
}
\value{
A formatted vector of messages
}
\description{
Format messages attribute for writing
}
