% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_metadata}
\alias{write_metadata}
\title{Format timber.log's metadata attributes for writing}
\usage{
write_metadata()
}
\value{
A vector of timber.log's metadata attributes
}
\description{
Format timber.log's metadata attributes for writing
}
\examples{
write_metadata()

}
