% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_session_info}
\alias{write_session_info}
\title{Format log.rx's session info attribute for writing}
\usage{
write_session_info()
}
\value{
A vector of log.rx's session info attribute
}
\description{
Format log.rx's session info attribute for writing
}
