% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_unapproved_functions}
\alias{write_unapproved_functions}
\title{Formats and returns a vector of unapproved functions}
\usage{
write_unapproved_functions()
}
\value{
Formatted vector of unapproved functions
}
\description{
\code{write_unapproved_functions()} gets log.rx unapproved_packages_functions
attribute, formats and returns the list of unapproved functions
}
\examples{
\dontrun{
write_unapproved_functions()
}
}
