% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_used_functions}
\alias{write_used_functions}
\title{Format used functions attribute for writing}
\usage{
write_used_functions()
}
\value{
A formatted vector of used functions
}
\description{
Format used functions attribute for writing
}
\examples{
\dontrun{
write_used_functions()
}

}
