% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_warnings}
\alias{write_warnings}
\title{Format warnings attribute for writing}
\usage{
write_warnings()
}
\value{
A formatted vector of warnings
}
\description{
Format warnings attribute for writing
}
\examples{
write_warnings()

}
