% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axecute.R
\name{axecute}
\alias{axecute}
\title{Create a log and run a file}
\usage{
axecute(
  file,
  log_name = NA,
  log_path = NA,
  remove_log_object = TRUE,
  quit_on_error = TRUE,
  to_report = c("messages", "output", "result")
)
}
\arguments{
\item{file}{File path of file to run}

\item{log_name}{Name of log file}

\item{log_path}{Path to output log to}

\item{remove_log_object}{Should the log object be removed after writing, defaults to TRUE}

\item{quit_on_error}{Should the session quit with status of 1 with error, defaults to TRUE}

\item{to_report}{toggle for optional reporting objects.
choices outlined below, may include as many as necessary
messages: any messages generated by program execution
output: any output generated by program execution
result: any result generated by program execution}
}
\value{
0 if there are no errors and 1 if any error
}
\description{
Create a log and run a file
}
