% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axecute.R
\name{axecute}
\alias{axecute}
\title{Create a log and run a file}
\usage{
axecute(file, log_name = NA, log_path = NA, remove_log_object = TRUE)
}
\arguments{
\item{file}{File path of file to run}

\item{log_name}{Name of log file}

\item{log_path}{Path to output log to}

\item{remove_log_object}{Should the log object be removed after writing, defaults to TRUE}
}
\value{
Nothing
}
\description{
Create a log and run a file
}
\examples{

scriptPath <- tempfile()
logDir <- tempdir()

writeLines("print('hello timber')", con = scriptPath)

axecute(scriptPath)
axecute(scriptPath, log_name = "log_out", log_path = logDir)

}
