% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-approved.R
\name{build_approved}
\alias{build_approved}
\title{Build approved packages and functions tibble}
\usage{
build_approved(pkg_list, file = NULL)
}
\arguments{
\item{pkg_list}{a named list of character vectors where the name is the
package name with a chacter vector of approved functions or 'All'}

\item{file}{(optional) name of file where the approved tibble will be read to.
If not specified, the tibble is returned.}
}
\value{
a tibble with two columns (library, function) and one row per function
}
\description{
A utility function to help you build your approve package and functions list.
This can be used by timber to log unnaproved use.
}
\details{
#' For more details see the help vignette:
\code{vignette("approved", package = "timber")}
}
\examples{
approved_pkgs <- list(
  base = c("library", "mean"),
  dplyr = "All"
)

# build and return
build_approved(approved_pkgs)

# build and save
dir <- tempdir()
build_approved(approved_pkgs, file.path(dir, "approved.rds"))
}
