% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{get_log_element}
\alias{get_log_element}
\title{Gets the value of a named element in the timber.log environment}
\usage{
get_log_element(el_key)
}
\arguments{
\item{el_key}{the key of the element in timber.log to be fetched}
}
\value{
Value of corresponding element from timber.log environment
}
\description{
Gets the value of a named element in the timber.log environment
}
\examples{
get_log_element("user")

}
