% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_cleanup}
\alias{log_cleanup}
\title{Cleaning-up of log.rx object}
\usage{
log_cleanup()
}
\value{
List of non-NA log.rx attributes and their values
}
\description{
\code{log_cleanup()} compiles a list of non-NA log.rx attributes and their
values, and return it
}
