% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_cleanup}
\alias{log_cleanup}
\title{Cleans up log and does checks against elements}
\usage{
log_cleanup()
}
\value{
List of non-NA elements and their value in timber.log environment
}
\description{
Cleans up log and does checks against elements
}
\examples{
scriptPath <- tempfile()
logDir <- tempdir()
writeLines("print('hello timber')", con = scriptPath)

log_remove()

log_init()

log_config(scriptPath)

log_cleanup()

log_write()

}
