% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_config}
\alias{log_config}
\title{Configuration of the log.rx environment}
\usage{
log_config(file = NA, log_name = NA, log_path = NA)
}
\arguments{
\item{file}{String. Path to file executed. Optional}

\item{log_name}{String. Name of log file. Optional}

\item{log_path}{String. Path to log file. Optional}
}
\value{
Nothing
}
\description{
\code{log_config()} initialises the log.rx environment, adds its attributes, and
sets them
}
