% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_config}
\alias{log_config}
\title{Configures the log.rx environment}
\usage{
log_config(file = NA, log_name = NA, log_path = NA)
}
\arguments{
\item{file}{File path of file being run, optional}

\item{log_name}{The log name}

\item{log_path}{The log path}
}
\value{
Nothing
}
\description{
Configures the log.rx environment
}
