% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_config}
\alias{log_config}
\title{Configures the timber.log environment}
\usage{
log_config(file = NA)
}
\arguments{
\item{file}{File path of file being run, optional}
}
\value{
Nothing
}
\description{
Configures the timber.log environment
}
\examples{
scriptPath <- tempfile()
logDir <- tempdir()
writeLines("print('hello timber')", con = scriptPath)

log_remove()

log_init()

log_config(scriptPath)

log_cleanup()

log_write()

}
