% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_config}
\alias{log_config}
\title{Configures the timber_log object within the timber.log environment}
\usage{
log_config()
}
\description{
Configures the timber_log object within the timber.log environment
}
\examples{
log_config()

}
