% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_config}
\alias{log_config}
\title{Configures the timber_log object}
\usage{
log_config()
}
\value{
timber_log object as a named list of attributes
}
\description{
Configures the timber_log object
}
\examples{
log_config()

}
