% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_init}
\alias{log_init}
\title{Initialises the timber_log object}
\usage{
log_init()
}
\value{
timber_log object as an empty named list
}
\description{
Initialises the timber_log object
}
\examples{
log_init()

}
