% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_init}
\alias{log_init}
\title{Initialization of the log.rx environment}
\usage{
log_init()
}
\value{
Nothing
}
\description{
\code{log_init()} initialises the log.rx environment
}
\examples{
# Initialise the log.rx environment
log_init()

# Remove the log.rx environment
log_remove()
}
