% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_remove}
\alias{log_remove}
\title{Remove the log.rx environment by setting options("log.rx") to NULL}
\usage{
log_remove()
}
\value{
Nothing
}
\description{
Remove the log.rx environment by setting options("log.rx") to NULL
}
