% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_remove}
\alias{log_remove}
\title{Remove the timber.log environment by setting options("timber.log") to NULL}
\usage{
log_remove()
}
\value{
Nothing
}
\description{
Remove the timber.log environment by setting options("timber.log") to NULL
}
\examples{
scriptPath <- tempfile()
logDir <- tempdir()
writeLines("print('hello timber')", con = scriptPath)

log_remove()

log_init()

log_config(scriptPath)

log_cleanup()

log_write()

}
