% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_write}
\alias{log_write}
\title{Write the formatted timber.log to a file}
\usage{
log_write()
}
\value{
Nothing
}
\description{
Write the formatted timber.log to a file
}
\examples{
scriptPath <- tempfile()
logDir <- tempdir()
writeLines("print('hello timber')", con = scriptPath)

log_remove()

log_init()

log_config(scriptPath)

log_cleanup()

log_write()

}
