% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_write}
\alias{log_write}
\title{Write the formatted timber.log to a file}
\usage{
log_write()
}
\arguments{
\item{log_name}{Log file name}

\item{log_path}{Log file path}
}
\value{
Nothing
}
\description{
Write the formatted timber.log to a file
}
\examples{
log_write()

}
