% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrxAddin.R
\name{logrxAddin}
\alias{logrxAddin}
\title{logrxAddin
Code needed to create the miniUI for the Addin}
\usage{
logrxAddin()
}
\value{
returns miniUI Addin to batch submit r files and create logs around them
}
\description{
logrxAddin
Code needed to create the miniUI for the Addin
}
