% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{run_safely_quietly}
\alias{run_safely_quietly}
\title{Safely run an R script and record results, outputs, messages, errors, warnings}
\usage{
run_safely_quietly(file_name)
}
\arguments{
\item{file_name}{Function to run}
}
\value{
Nothing
}
\description{
Safely run an R script and record results, outputs, messages, errors, warnings
}
