% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{set_log_element}
\alias{set_log_element}
\title{Adds values to existing named elements in the timber_log object}
\usage{
set_log_element(el_key, el_value)
}
\arguments{
\item{el_key}{the key of the element in timber_log to be updated}

\item{el_value}{the value to be added to the timber_log element}
}
\value{

}
\description{
Adds values to existing named elements in the timber_log object
}
\examples{
set_log_element("user", Sys.info()[["user"]])
}
