% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{set_log_element}
\alias{set_log_element}
\title{Adds values to existing named elements in the timber.log environment}
\usage{
set_log_element(el_key, el_value)
}
\arguments{
\item{el_key}{the key of the element in timber.log to be updated}

\item{el_value}{the value to be added to the timber.log element}
}
\value{
Nothing
}
\description{
Adds values to existing named elements in the timber.log environment
}
\examples{
set_log_element("user", Sys.info()[["user"]])

}
