% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{set_log_name_path}
\alias{set_log_name_path}
\title{Set the log name and path:
\enumerate{
\item As the name and path if specified
\item As the file name and path if specified
\item As timber_log.log and . if none of the above are specified
}}
\usage{
set_log_name_path(log_name = NA, log_path = NA)
}
\arguments{
\item{log_name}{The log name}

\item{log_path}{The log path}
}
\value{
Nothing
}
\description{
Set the log name and path:
\enumerate{
\item As the name and path if specified
\item As the file name and path if specified
\item As timber_log.log and . if none of the above are specified
}
}
