% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_log_element}
\alias{write_log_element}
\title{Generic function to format log.rx elements for writing}
\usage{
write_log_element(el_key, prefix = NULL)
}
\arguments{
\item{el_key}{the key of the element in log.rx to be fetched}

\item{prefix}{string to be placed before element value during formatting}
}
\value{
formatted element including prefix
}
\description{
Generic function to format log.rx elements for writing
}
