% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_log_header}
\alias{write_log_header}
\title{Generic function to format log section headers}
\usage{
write_log_header(title_string)
}
\arguments{
\item{title_string}{String to be used as section title}
}
\value{
A vector with the header including title
}
\description{
Generic function to format log section headers
}
