% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writers.R
\name{write_metadata}
\alias{write_metadata}
\title{Format timer.log's metadata attributes for writing}
\usage{
write_metadata()
}
\value{
A vector of timer.log's metadata attributes
}
\description{
Format timer.log's metadata attributes for writing
}
\examples{
write_metadata()

}
