% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_metadata}
\alias{write_metadata}
\title{Format log.rx's metadata attributes for writing}
\usage{
write_metadata()
}
\value{
A vector of log.rx's metadata attributes
}
\description{
Format log.rx's metadata attributes for writing
}
