% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_unapproved_functions}
\alias{write_unapproved_functions}
\title{Format unapproved functions attribute for writing}
\usage{
write_unapproved_functions()
}
\value{
A formatted vector of unapproved functions
}
\description{
Format unapproved functions attribute for writing
}
\examples{
\dontrun{
write_unapproved_functions()
}
}
