% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{write_used_functions}
\alias{write_used_functions}
\title{Formats and returns a vector of used package functions}
\usage{
write_used_functions()
}
\value{
Formatted vector of used package functions
}
\description{
\code{write_used_functions()} gets log.rx used_packages_functions attribute,
formats and returns the list of used package functions
}
\examples{
\dontrun{
write_used_functions()
}

}
